+++ frb_example/pure_dart/dart/lib/bridge_generated.dart
@@ -6,6 +6,7 @@
 import 'dart:convert';
 import 'dart:typed_data';
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:meta/meta.dart' as meta;
 
 import 'dart:convert';
 import 'dart:typed_data';
@@ -100,6 +101,8 @@ abstract class FlutterRustBridgeExample {
   Future<List<Point>> getComplexArray({dynamic hint});
 
   Future<int> getUsize({required int u, dynamic hint});
+
+  Future<UserId> nextUserId({required UserId userId, dynamic hint});
 }
 
 class ApplicationEnv {
@@ -309,6 +312,15 @@ class Point {
     required this.x,
     required this.y,
   });
+}
+
+/// Example for @freezed and @meta.immutable
+@freezed
+@meta.immutable
+class UserId with _$UserId {
+  const factory UserId({
+    required int value,
+  }) = _UserId;
 }
 
 class VecOfPrimitivePack {
@@ -798,6 +810,17 @@ class FlutterRustBridgeExampleImpl extends FlutterRustBridgeBase<FlutterRustBrid
         hint: hint,
       ));
 
+  Future<UserId> nextUserId({required UserId userId, dynamic hint}) => executeNormal(FlutterRustBridgeTask(
+        callFfi: (port_) => inner.wire_next_user_id(port_, _api2wire_box_autoadd_user_id(userId)),
+        parseSuccessData: _wire2api_user_id,
+        constMeta: const FlutterRustBridgeTaskConstMeta(
+          debugName: "next_user_id",
+          argNames: ["userId"],
+        ),
+        argValues: [userId],
+        hint: hint,
+      ));
+
   // Section: api2wire
   ffi.Pointer<wire_uint_8_list> _api2wire_String(String raw) {
     return _api2wire_uint_8_list(utf8.encoder.convert(raw));
@@ -896,6 +919,12 @@ class FlutterRustBridgeExampleImpl extends FlutterRustBridgeBase<FlutterRustBrid
   ffi.Pointer<wire_NewTypeInt> _api2wire_box_autoadd_new_type_int(NewTypeInt raw) {
     final ptr = inner.new_box_autoadd_new_type_int();
     _api_fill_to_wire_new_type_int(raw, ptr.ref);
+    return ptr;
+  }
+
+  ffi.Pointer<wire_UserId> _api2wire_box_autoadd_user_id(UserId raw) {
+    final ptr = inner.new_box_autoadd_user_id();
+    _api_fill_to_wire_user_id(raw, ptr.ref);
     return ptr;
   }
 
@@ -1218,6 +1247,10 @@ class FlutterRustBridgeExampleImpl extends FlutterRustBridgeBase<FlutterRustBrid
     _api_fill_to_wire_new_type_int(apiObj, wireObj.ref);
   }
 
+  void _api_fill_to_wire_box_autoadd_user_id(UserId apiObj, ffi.Pointer<wire_UserId> wireObj) {
+    _api_fill_to_wire_user_id(apiObj, wireObj.ref);
+  }
+
   void _api_fill_to_wire_box_exotic_optionals(ExoticOptionals apiObj, ffi.Pointer<wire_ExoticOptionals> wireObj) {
     _api_fill_to_wire_exotic_optionals(apiObj, wireObj.ref);
   }
@@ -1324,6 +1357,10 @@ class FlutterRustBridgeExampleImpl extends FlutterRustBridgeBase<FlutterRustBrid
 
   void _api_fill_to_wire_opt_box_exotic_optionals(ExoticOptionals? apiObj, ffi.Pointer<wire_ExoticOptionals> wireObj) {
     if (apiObj != null) _api_fill_to_wire_box_exotic_optionals(apiObj, wireObj);
+  }
+
+  void _api_fill_to_wire_user_id(UserId apiObj, wire_UserId wireObj) {
+    wireObj.value = _api2wire_u32(apiObj.value);
   }
 }
 
@@ -1784,6 +1821,14 @@ Uint8List _wire2api_uint_8_list(dynamic raw) {
 
 void _wire2api_unit(dynamic raw) {
   return;
+}
+
+UserId _wire2api_user_id(dynamic raw) {
+  final arr = raw as List<dynamic>;
+  if (arr.length != 1) throw Exception('unexpected arr length: expect 1 but see ${arr.length}');
+  return UserId(
+    value: _wire2api_u32(arr[0]),
+  );
 }
 
 int _wire2api_usize(dynamic raw) {
@@ -2381,6 +2426,20 @@ class FlutterRustBridgeExampleWire implements FlutterRustBridgeWireBase {
       _lookup<ffi.NativeFunction<ffi.Void Function(ffi.Int64, uintptr_t)>>('wire_get_usize');
   late final _wire_get_usize = _wire_get_usizePtr.asFunction<void Function(int, int)>();
 
+  void wire_next_user_id(
+    int port_,
+    ffi.Pointer<wire_UserId> user_id,
+  ) {
+    return _wire_next_user_id(
+      port_,
+      user_id,
+    );
+  }
+
+  late final _wire_next_user_idPtr =
+      _lookup<ffi.NativeFunction<ffi.Void Function(ffi.Int64, ffi.Pointer<wire_UserId>)>>('wire_next_user_id');
+  late final _wire_next_user_id = _wire_next_user_idPtr.asFunction<void Function(int, ffi.Pointer<wire_UserId>)>();
+
   ffi.Pointer<wire_StringList> new_StringList(
     int len,
   ) {
@@ -2530,6 +2589,14 @@ class FlutterRustBridgeExampleWire implements FlutterRustBridgeWireBase {
       _lookup<ffi.NativeFunction<ffi.Pointer<wire_NewTypeInt> Function()>>('new_box_autoadd_new_type_int');
   late final _new_box_autoadd_new_type_int =
       _new_box_autoadd_new_type_intPtr.asFunction<ffi.Pointer<wire_NewTypeInt> Function()>();
+
+  ffi.Pointer<wire_UserId> new_box_autoadd_user_id() {
+    return _new_box_autoadd_user_id();
+  }
+
+  late final _new_box_autoadd_user_idPtr =
+      _lookup<ffi.NativeFunction<ffi.Pointer<wire_UserId> Function()>>('new_box_autoadd_user_id');
+  late final _new_box_autoadd_user_id = _new_box_autoadd_user_idPtr.asFunction<ffi.Pointer<wire_UserId> Function()>();
 
   ffi.Pointer<ffi.Uint8> new_box_bool(
     bool value,
@@ -3063,6 +3130,11 @@ class wire_ApplicationSettings extends ffi.Struct {
   external int mode;
 
   external ffi.Pointer<wire_ApplicationEnv> env;
+}
+
+class wire_UserId extends ffi.Struct {
+  @ffi.Uint32()
+  external int value;
 }
 
 typedef uintptr_t = ffi.UnsignedLong;